Changed engine to allow for multi-board support
The previous interface only allowed for 2 boards, setup in a top-bottom setup.
The interface now includes 2 methods to start (StartEngine and StartEngine_TopBottom). The first one can be tunned to use OpenMPD with an arbitrary number of boards (and arbitrary arrangement).
Minor changes to the vertical speed test. It still fails some times when used with force sync...