Abstracted Primitive into C# interface and PrimitiveGameObject
This allows for PrimitiveGameObjects and library extensions or user code such as PrimitiveNode (from the graph design tool) to register themselves with the OpenMPD_PresentationManager in a uniform way. I'm committing this separately from the node design code to allow for easy merging with the main branch if desired.

(C# does not allow for multiple inheritance and so this seemed the best way to group similiar objects